pyproject.toml
[project]
name = "sesoko"
version = "0.1.0"
description = "Image captioning and LoRA training for martial arts domain"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    # Core ML
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "accelerate>=1.12.0",
    "pillow>=12.1.0",
    "pillow-heif>=1.1.0",
    "bitsandbytes>=0.49.0",
    "sentencepiece>=0.2.0",
    "protobuf>=6.33.1",
    
    # Captioning
    "matplotlib>=3.10.7",
    "tomli-w>=1.1.0",
    
    # LoRA Training
    "click>=8.3.1",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
    "peft>=0.18.0",
    "safetensors>=0.4.0",
    "ultralytics>=8.3.235",
]

# PyTorch index for CUDA 13.0
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--verbose --cov=src --cov-report=term-missing --cov-report=html"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint.isort]
known-first-party = ["caption_images"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
